refactor(chat): split memory forget tests
